[PWGCF] Update femto framework (#13568)

+++ PWGCF/Femto/Core/dataTypes.h
@@ -29,7 +29,7 @@ using CollisionTagType = uint64_t;
 using CollisionMaskType = uint16_t;
 
 // datatypes for tracks
-using MomentumType = uint16_t;
+using MomentumType = uint8_t;
 using TrackMaskType = uint64_t;
 using TrackType = uint16_t;
